backend/requirements.txt
Flask>=2.0
requests>=2.28
python-dotenv>=0.21
Flask-Cors>=3.0
flask
flask-cors
python-dotenv
langchain
langchain-openai
langchain-community
tavily-python
numpy
opencv-python